src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/reference/ArrayFloatAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.attribute.ArrayAttributeParser;
import org.openqa.selenium.WebElement;

public class ArrayFloatAttributeParser extends ArrayAttributeParser {

    protected ArrayFloatAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Float.parseFloat(element.getAttribute(super.attribute));
    }

}
